requirements.txt
flask
redis
tqdm
requests
mysql-connector-python